refactor: ResearchFieldCollection
