+++ frontend/components/home/helmholtz/index.tsx
@@ -27,6 +27,7 @@ import logger from '~/utils/logger'
 import Image from 'next/image'
 import ParticipatingOrganisations from '~/components/home/helmholtz/ParticipatingOrganisations'
 import LatestSpotlight from '~/components/home/helmholtz/LatestSpotlight'
+import ResearchFieldCollection from './ResearchFieldCollection'
 /*! purgecss end ignore */
 
 type SpotlightDescription = {
@@ -35,15 +36,6 @@ type SpotlightDescription = {
   image: string,
   link: string
 }
-
-const RESEARCH_FIELDS = [
-  {key: 1, name: 'Energy', img: '/images/pexels-pixabay-414837.jpg'},
-  {key: 2, name: 'Earth & Environment', img: '/images/pexels-blue-ox-studio-695299.jpg'},
-  {key: 3, name: 'Health', img: '/images/pexels-rfstudio-3825529.jpg'},
-  {key: 4, name: 'Information', img: '/images/jj-ying-8bghKxNU1j0-unsplash.jpg'},
-  {key: 5, name: 'Aeronautics, Space and Transport', img: '/images/pexels-aleksejs-bergmanis-681335.jpg'},
-  {key: 6, name: 'Matter', img: '/images/desy_yulia-buchatskaya-hYvZHggmuc4-unsplash.jpg'},
-]
 
 const SPOTLIGHTS= [
   {
@@ -152,38 +144,6 @@ function Spotlights({spotlights}:{spotlights: Array<SpotlightDescription>}) {
           )
         })}
       </div> */}
-    </div>
-  )
-}
-
-function ResearchField({background, name}:{background: string, name: string}) {
-
-  function mouseEnter(event: React.MouseEvent<HTMLAnchorElement>) {
-    if (!(event.target instanceof HTMLAnchorElement)) return
-    const background = event.target.dataset.background
-    event.target.parentElement!.parentElement!.style.backgroundImage = 'url("' + background + '")'
-  }
-
-  const uriComponent = `["${name}"]`
-  const link=`/software?&keywords=${encodeURIComponent(uriComponent)}&page=0&rows=12`
-
-  return (
-    <a className='underline hover:text-white' onMouseEnter={mouseEnter} data-background={background} href={link}>
-      {name}
-    </a>
-  )
-}
-
-function ResearchFields() {
-  return (
-    <div id="researchTopicBox" className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-x-10 gap-y-10 md:gap-y-20 text-center text-3xl place-items-center py-16">
-      {
-        RESEARCH_FIELDS.map(item => {
-          return(
-            <ResearchField key={`researchfield-${item.key}`} background={item.img} name={item.name} />
-          )
-        })
-      }
     </div>
   )
 }
@@ -312,7 +272,7 @@ export default function Home() {
             onMouseLeave={resetBackgroundImage}>
             <h2 className='text-5xl'>Discover software by research topic</h2>
             {/* <div className="text-xl my-4">Browse Software by Research Topic</div> */}
-            <ResearchFields />
+            <ResearchFieldCollection />
           </div>
         </div>
